ADDED SCHEDULE FN() TO FIND RANK FOR GIVEN SEASON WEEK TEAM COMBO
In a two-birds-one-stone effort, a new function has been created as part of Schedule()
which can be used to return rank information for a given season, week, and team.
Eventually this will be used to not only improve readability (bird 1), but also
to add rank information to a couple of emailed frames which currently don't
include the rank information.