ADDED DB INGESTION AND TABLE DUMPS TO CSV I am transitioning to a database-drive approach so I can eventually work on Elo ratings and other predictive capabilities.  As a result, I am in the midst of a rewrite utilizing a local database. I will dump my database tables to the ./csv/ folder so users of this repo can ingest the data into dataframes or their own database.
Note that this is still largely a work in progress.

Additionally, note that the cfb_data.py script is currently not working.  I suspect it's because
the playoff rankings came out and are causing problems with the dataset becuase their week
numbers probably don't match up with the schedule week numbers.  This may or may not be the
issue, and fixing will require some time.