CODE WORKING AGAIN - FIXED BLOCKING ISSUE
When the code was written, I failed to consider a scenario where two teams would play each
other for the first time.  Oddly, this scenario is occurring between two conference opponents
when Georgia plays Texas A&M.  Because they haven't played eachother before (or at least
not within the num_past_yr range set in my config.ini) it means that the series record
is incalculable.  I had to write some if/then code to handle for this, and the code
is functional now once again.